Fix for broken update of Transaction object after update
TransactionScopeProxy implementation broke the update of the transaction
object, although it was initially broken in a different way (because of
threading).

Also changed the unit test REST harness, so that the response is actually
calculated in another thread, instead of being precalculated before the
thread actually started (which alters the expected behavior for this bug).