Detects when a MemberExpression returns null and adjusts it as a ConstantExpression
Neo4j doesn't consider null as valid values for PropertyBags
(relationships and nodes).
Neo4jClient could have a MemberExpression in a Where() construct and it
could detect it if was in reality a ConstantExpression, however it just
sends the value as parameter regardless if it's a null value.

This commit fixes that by having the same behavior that occurs in
VisitConstant(ConstantExpression) if it detects a null value.