Make sure culture is set before deserializer tests
Looks like the Class fixture is run once for every test, but the tests
themselves is run in separate threads. The entire
IClassFixture<CultureInfoSetupFixture> pattern thus looks wrong. This
commit will use a ctor instead as that is code which runs before each
test, thus making sure the thread which runs the tests gets a
deterministic culture.

While this solves the problem for our cases, the entire pattern should
be revisited to work correctly.

PS: This might be specific to running tests through ReSharper, but would
still be useful to fix as many people use ReSharper.