Workaround for .NET 4.7.1 "Transaction.Current not restored" bug
Ref https://github.com/Microsoft/dotnet-framework-early-access/issues/7

After a call crossing AppDomain boundaries, Transaction.Current would be null.
This causes us to loose the current transaction when a neo4j transaction gets
promoted to a distributed transaction.

This workaround of manually restoring the transaction resolves this issue.
While not tested on earlier versions of .NET, it *hopefully* won't cause any
problems.

The biggest concern is Jim Carleys comment stating
> The fix I implemented in 4.7.2 for this issue has bad side effects that
> break other remoting scenarios and those problems cannot be worked around.
Ref https://github.com/Microsoft/dotnet-framework-early-access/issues/7#issuecomment-415899496

In other words, this is a Works On My Machine fix.